@@ -5,23 +5,25 @@
   "dependencies": {
     "@emotion/react": "^11.10.0",
     "@emotion/styled": "^11.10.0",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/lab": "^5.0.0-alpha.94",
-    "@mui/material": "^5.10.0",
-    "@mui/styles": "^5.9.3",
-    "@tanstack/react-query": "^4.1.3",
-    "@tanstack/react-query-devtools": "^4.0.10",
+    "@mui/icons-material": "^5.10.2",
+    "@mui/lab": "^5.0.0-alpha.96",
+    "@mui/material": "^5.10.2",
+    "@mui/styles": "^5.10.2",
+    "@tanstack/react-query": "^4.2.3",
+    "@tanstack/react-query-devtools": "^4.2.3",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
-    "aws-amplify": "^4.3.31",
+    "aws-amplify": "^4.3.33",
     "axios": "^0.21.4",
     "bootstrap": "^5.1.3",
     "dotenv": "^10.0.0",
     "js-cookie": "^3.0.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-bootstrap": "^2.5.0",
-    "react-dom": "^17.0.2",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
+    "react-dom": "^18.2.0",
     "react-router": "^5.3.3",
     "react-router-dom": "^5.3.3",
     "react-scripts": "^5.0.1",
